refactor: simplify module structure
Cleaned up project organization by removing unused code and consolidating small related modules.

**Removed:**
- Unused error.rs, event_listener.rs, and query_builder.rs modules

**Consolidated:**
- Merged command.rs, player.rs, and pug.rs into types.rs (154 lines)
- Renamed data.rs to constants.rs for clarity

**Updated:**
- All import paths across handlers, database, and models
- Module declarations in lib.rs and main.rs

Reduces file count from 24 to 21 Rust files while maintaining all functionality.
