--- src/event_listener.rs
@@ -1,9 +1,0 @@
-use std::sync::Arc;
-
-pub struct EventListener {}
-
-impl EventListener {
-    pub fn new(db: Arc<Database>) -> Self {
-        Self { db }
-    }
-}
