+++ src/lib.rs
@@ -3,7 +3,6 @@
 
 // Library crate for modules
 pub mod database;
-pub mod error;
 pub mod handlers;
 pub mod models;
