+++ src/models/server.rs
@@ -10,7 +10,7 @@ use tracing::{info, warn};
 use anyhow::{anyhow, Error, Result};
 
 use crate::handlers::player::check_role;
-use crate::models::data::*;
+use crate::models::constants::*;
 use crate::models::session::*;
 use crate::models::{CommandContext, ComponentContext};
